requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
stripe==7.8.0
requests==2.31.0
qrcode==7.4.2
python-dotenv==1.0.0
